Rename cacheFirstDataSourceModule.ts to cache.first.data.source.module.ts
