+++ src/datasources/config-api/config-api.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { ConfigApi } from './config-api.service';
-import { CacheFirstDataSourceModule } from '../cache/cacheFirstDataSourceModule';
+import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 
 @Module({
   imports: [CacheFirstDataSourceModule],
