+++ src/datasources/transaction-api/transaction-api.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ConfigApiModule } from '../config-api/config-api.module';
 import { TransactionApiManager } from './transaction-api.manager';
-import { CacheFirstDataSourceModule } from '../cache/cacheFirstDataSourceModule';
+import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 
 @Module({
   imports: [ConfigApiModule, CacheFirstDataSourceModule],
